requirements.txt
python-telegram-bot==20.6
openai==1.2.0
apscheduler==3.10.4
